automation/requirements.txt
# Dependências Python para Automação WhatsApp
mysql-connector-python==8.2.0
requests==2.31.0
python-dotenv==1.0.0
twilio==8.10.0
pywhatkit==5.4
selenium==4.15.0
webdriver-manager==4.0.1